backend/resxiv_backend/requirements_production.txt
# Production Requirements - L6 Engineering Standards
# Reduced from 162 to 35 essential packages (78% reduction)

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# MongoDB for unstructured data
motor==3.3.2
pymongo==4.6.0

# Redis for caching
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI/ML Core - Production Grade
langchain-openai==0.0.5
langgraph==0.0.26

# HTTP Clients
httpx==0.25.2
aiofiles==23.2.1

# Configuration & Environment
python-dotenv==1.0.0

# Email & Templates
aiosmtplib==3.0.1
jinja2==3.1.2

# File Processing (Essential Only)
PyPDF2==3.0.1

# Git Integration
pygit2>=1.14.0

# Testing (development/CI only)
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring & Logging
structlog==23.2.0

# Rate Limiting
slowapi==0.1.9

# Date/Time
python-dateutil==2.8.2

# Validation & Email
email-validator>=2.1.1
pydantic[email]==2.5.0 